requirements.txt
flask
werkzeug
pandas
yfinance
numpy
Pillow
psycopg2-binary
gunicorn
gunicorn
psycopg2-binary
